Move the printing code from MainClass to a new ObjectPrinter class.
