+++ ObjectPrinter/ObjectPrinter.csproj
@@ -39,6 +39,7 @@
     <Compile Include="TypeHelper.cs" />
     <Compile Include="Collector.cs" />
     <Compile Include="FbxVisitor.cs" />
+    <Compile Include="ObjectPrinter.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
